Fix stats aggregations day comparison (#35)
`time.Day()` returns "the day of the month" and the stats aggregator
used this as a check if `time.Now().Day()` has increased from the last
iterations `time.Now().Day()`. Depending on which day of month
the service was started, this leads to situation where stats are only
updated in the end of the month.

This change adds `monotonicDay()` which computes number of days since
[Epoch](https://en.wikipedia.org/wiki/Unix_time) for given `time.Time`
and fixes the comparison.